+++ master/src/main/java/org/apache/celeborn/service/deploy/master/slotsalloc/SlotsAllocator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.service.deploy.master;
+package org.apache.celeborn.service.deploy.master.slotsalloc;
 
 import static org.apache.celeborn.common.protocol.StorageInfo.Type.*;
 
@@ -26,163 +26,37 @@ import java.util.stream.Collectors;
 import scala.Tuple2;
 import scala.Tuple3;
 
-import org.apache.commons.lang3.StringUtils;
 import org.roaringbitmap.RoaringBitmap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.apache.celeborn.common.meta.DiskInfo;
-import org.apache.celeborn.common.meta.DiskStatus;
 import org.apache.celeborn.common.meta.WorkerInfo;
 import org.apache.celeborn.common.protocol.PartitionLocation;
 import org.apache.celeborn.common.protocol.StorageInfo;
 
 public class SlotsAllocator {
-  static class UsableDiskInfo {
-    DiskInfo diskInfo;
-    long usableSlots;
-
-    /** @param diskInfo will be used as source for usableSlots. */
-    UsableDiskInfo(DiskInfo diskInfo) {
-      this.diskInfo = diskInfo;
-      this.usableSlots = diskInfo.getAvailableSlots();
-    }
-
-    UsableDiskInfo(DiskInfo diskInfo, long usableSlots) {
-      this.diskInfo = diskInfo;
-      this.usableSlots = usableSlots;
-    }
-  }
-
   private static final Logger logger = LoggerFactory.getLogger(SlotsAllocator.class);
   private static final Random rand = new Random();
-  private static boolean initialized = false;
-  private static double[] taskAllocationRatio = null;
 
   public static Map<WorkerInfo, Tuple2<List<PartitionLocation>, List<PartitionLocation>>>
-      offerSlotsRoundRobin(
+      offerSlots(
           List<WorkerInfo> workers,
           List<Integer> partitionIds,
           boolean shouldReplicate,
           boolean shouldRackAware,
           int availableStorageTypes,
           boolean interruptionAware,
-          int interruptionAwareThreshold) {
+          int interruptionAwareThreshold,
+          SlotsAssignStrategy strategy) {
     if (partitionIds.isEmpty()) {
       return new HashMap<>();
     }
     if (workers.size() < 2 && shouldReplicate) {
       return new HashMap<>();
     }
-    Map<WorkerInfo, List<UsableDiskInfo>> slotsRestrictions = new HashMap<>();
-    for (WorkerInfo worker : workers) {
-      List<UsableDiskInfo> usableDisks =
-          slotsRestrictions.computeIfAbsent(worker, v -> new ArrayList<>());
-      for (DiskInfo diskInfo : worker.diskInfos().values()) {
-        if (DiskStatus.HEALTHY.equals(diskInfo.status())
-            && StorageInfo.isAvailable(diskInfo.storageType(), availableStorageTypes)) {
-          usableDisks.add(new UsableDiskInfo(diskInfo));
-        }
-      }
-    }
-    return locateSlots(
-        partitionIds,
-        workers,
-        slotsRestrictions,
-        shouldReplicate,
-        shouldRackAware,
-        availableStorageTypes,
-        interruptionAware,
-        interruptionAwareThreshold);
-  }
-
-  /**
-   * It assumes that all disks whose available space is greater than the minimum space are divided
-   * into multiple groups. A faster group will allocate more allocations than a slower group by
-   * diskGroupGradient.
-   */
-  public static Map<WorkerInfo, Tuple2<List<PartitionLocation>, List<PartitionLocation>>>
-      offerSlotsLoadAware(
-          List<WorkerInfo> workers,
-          List<Integer> partitionIds,
-          boolean shouldReplicate,
-          boolean shouldRackAware,
-          int diskGroupCount,
-          double diskGroupGradient,
-          double flushTimeWeight,
-          double fetchTimeWeight,
-          double activeSlotsWeight,
-          int availableStorageTypes,
-          boolean interruptionAware,
-          int interruptionAwareThreshold) {
-    if (partitionIds.isEmpty()) {
-      return new HashMap<>();
-    }
-    if (workers.size() < 2 && shouldReplicate) {
-      return new HashMap<>();
-    }
-    if (StorageInfo.HDFSOnly(availableStorageTypes)
-        || StorageInfo.S3Only(availableStorageTypes)
-        || StorageInfo.OSSOnly(availableStorageTypes)) {
-      return offerSlotsRoundRobin(
-          workers,
-          partitionIds,
-          shouldReplicate,
-          shouldRackAware,
-          availableStorageTypes,
-          interruptionAware,
-          interruptionAwareThreshold);
-    }
-
-    List<DiskInfo> usableDisks = new ArrayList<>();
-    Map<DiskInfo, WorkerInfo> diskToWorkerMap = new HashMap<>();
-
-    workers.forEach(
-        i ->
-            i.diskInfos()
-                .forEach(
-                    (key, diskInfo) -> {
-                      diskToWorkerMap.put(diskInfo, i);
-                      if (diskInfo.actualUsableSpace() > 0
-                          && diskInfo.status().equals(DiskStatus.HEALTHY)
-                          && !diskInfo.storageType().isDFS()) {
-                        usableDisks.add(diskInfo);
-                      }
-                    }));
-
-    boolean noUsableDisks =
-        usableDisks.isEmpty()
-            || (shouldReplicate
-                && (usableDisks.size() == 1
-                    || usableDisks.stream().map(diskToWorkerMap::get).distinct().count() <= 1));
-    boolean noAvailableSlots =
-        usableDisks.stream().mapToLong(DiskInfo::getAvailableSlots).sum() <= 0;
-
-    if (noUsableDisks || noAvailableSlots) {
-      logger.warn(
-          "offer slots for {} fallback to roundrobin because there is no {}",
-          StringUtils.join(partitionIds, ','),
-          noUsableDisks ? "usable disks" : "available slots");
-      return offerSlotsRoundRobin(
-          workers,
-          partitionIds,
-          shouldReplicate,
-          shouldRackAware,
-          availableStorageTypes,
-          interruptionAware,
-          interruptionAwareThreshold);
-    }
-
-    if (!initialized) {
-      initLoadAwareAlgorithm(diskGroupCount, diskGroupGradient);
-    }
-
     Map<WorkerInfo, List<UsableDiskInfo>> slotsRestrictions =
-        getSlotsRestrictionsByLoadAwareAlgorithm(
-            placeDisksToGroups(
-                usableDisks, diskGroupCount, flushTimeWeight, fetchTimeWeight, activeSlotsWeight),
-            diskToWorkerMap,
-            shouldReplicate ? partitionIds.size() * 2 : partitionIds.size());
+        strategy.computeSlotBudgets(workers, partitionIds, shouldReplicate, availableStorageTypes);
     return locateSlots(
         partitionIds,
         workers,
@@ -643,167 +517,6 @@ public class SlotsAllocator {
         || !Objects.equals(
             primaryWorkers.get(primaryIndex).networkLocation(),
             replicaWorkers.get(nextReplicaInd).networkLocation());
-  }
-
-  private static void initLoadAwareAlgorithm(int diskGroups, double diskGroupGradient) {
-    taskAllocationRatio = new double[diskGroups];
-    double totalAllocations = 0;
-
-    for (int i = 0; i < diskGroups; i++) {
-      totalAllocations += Math.pow(1 + diskGroupGradient, diskGroups - 1 - i);
-    }
-    for (int i = 0; i < diskGroups; i++) {
-      taskAllocationRatio[i] =
-          Math.pow(1 + diskGroupGradient, diskGroups - 1 - i) / totalAllocations;
-    }
-    logger.info(
-        "load-aware offer slots algorithm init with taskAllocationRatio {}",
-        StringUtils.join(taskAllocationRatio, ','));
-    initialized = true;
-  }
-
-  private static List<List<DiskInfo>> placeDisksToGroups(
-      List<DiskInfo> usableDisks,
-      int diskGroupCount,
-      double flushTimeWeight,
-      double fetchTimeWeight,
-      double activeSlotsWeight) {
-    List<List<DiskInfo>> diskGroups = new ArrayList<>();
-    usableDisks.sort(
-        (o1, o2) -> {
-          double delta =
-              (o1.avgFlushTime() * flushTimeWeight + o1.avgFetchTime() * fetchTimeWeight)
-                  + o1.activeSlots() * activeSlotsWeight
-                  - (o2.avgFlushTime() * flushTimeWeight
-                      + o2.avgFetchTime() * fetchTimeWeight
-                      + o2.activeSlots() * activeSlotsWeight);
-          return delta < 0 ? -1 : (delta > 0 ? 1 : 0);
-        });
-    int diskCount = usableDisks.size();
-    int startIndex = 0;
-    int groupSizeSize = (int) Math.ceil(usableDisks.size() / (double) diskGroupCount);
-    for (int i = 0; i < diskGroupCount; i++) {
-      List<DiskInfo> diskList = new ArrayList<>();
-      if (startIndex >= usableDisks.size()) {
-        continue;
-      }
-      if (startIndex + groupSizeSize <= diskCount) {
-        diskList.addAll(usableDisks.subList(startIndex, startIndex + groupSizeSize));
-        startIndex += groupSizeSize;
-      } else {
-        diskList.addAll(usableDisks.subList(startIndex, usableDisks.size()));
-        startIndex = usableDisks.size();
-      }
-      diskGroups.add(diskList);
-    }
-    return diskGroups;
-  }
-
-  /**
-   * This method implement the load aware slots allocation algorithm. See details at
-   * /docs/developers/slotsallocation.md
-   */
-  private static Map<WorkerInfo, List<UsableDiskInfo>> getSlotsRestrictionsByLoadAwareAlgorithm(
-      List<List<DiskInfo>> groups, Map<DiskInfo, WorkerInfo> diskWorkerMap, int partitionCnt) {
-    int groupSize = groups.size();
-    long[] groupAllocations = new long[groupSize];
-    Map<WorkerInfo, List<UsableDiskInfo>> restrictions = new HashMap<>();
-    long[] groupAvailableSlots = new long[groupSize];
-    for (int i = 0; i < groupSize; i++) {
-      for (DiskInfo disk : groups.get(i)) {
-        groupAvailableSlots[i] += disk.getAvailableSlots();
-      }
-    }
-    double[] currentAllocation = new double[groupSize];
-    double currentAllocationSum = 0;
-    for (int i = 0; i < groupSize; i++) {
-      if (!groups.get(i).isEmpty()) {
-        currentAllocationSum += taskAllocationRatio[i];
-      }
-    }
-    for (int i = 0; i < groupSize; i++) {
-      if (!groups.get(i).isEmpty()) {
-        currentAllocation[i] = taskAllocationRatio[i] / currentAllocationSum;
-      }
-    }
-    long toNextGroup = 0;
-    long left = partitionCnt;
-    for (int i = 0; i < groupSize; i++) {
-      if (left <= 0) {
-        break;
-      }
-      long estimateAllocation = (int) Math.ceil(partitionCnt * currentAllocation[i]);
-      if (estimateAllocation > left) {
-        estimateAllocation = left;
-      }
-      if (estimateAllocation + toNextGroup > groupAvailableSlots[i]) {
-        groupAllocations[i] = groupAvailableSlots[i];
-        toNextGroup = estimateAllocation - groupAvailableSlots[i] + toNextGroup;
-      } else {
-        groupAllocations[i] = estimateAllocation + toNextGroup;
-      }
-      left -= groupAllocations[i];
-    }
-
-    long groupLeft = 0;
-    for (int i = 0; i < groups.size(); i++) {
-      int disksInsideGroup = groups.get(i).size();
-      long groupRequired = groupAllocations[i] + groupLeft;
-      for (DiskInfo disk : groups.get(i)) {
-        if (groupRequired <= 0) {
-          break;
-        }
-        List<UsableDiskInfo> diskAllocation =
-            restrictions.computeIfAbsent(diskWorkerMap.get(disk), v -> new ArrayList<>());
-        long allocated =
-            (int) Math.ceil((groupAllocations[i] + groupLeft) / (double) disksInsideGroup);
-        if (allocated > disk.getAvailableSlots()) {
-          allocated = disk.getAvailableSlots();
-        }
-        if (allocated > groupRequired) {
-          allocated = groupRequired;
-        }
-        diskAllocation.add(new UsableDiskInfo(disk, Math.toIntExact(allocated)));
-        groupRequired -= allocated;
-      }
-      groupLeft = groupRequired;
-    }
-
-    if (logger.isDebugEnabled()) {
-      StringBuilder sb = new StringBuilder();
-      for (int i = 0; i < groups.size(); i++) {
-        sb.append("| group ").append(i).append(" ");
-        for (DiskInfo diskInfo : groups.get(i)) {
-          WorkerInfo workerInfo = diskWorkerMap.get(diskInfo);
-          String workerHost = workerInfo.host();
-          long allocation = 0;
-          if (restrictions.get(workerInfo) != null) {
-            for (UsableDiskInfo usableInfo : restrictions.get(workerInfo)) {
-              if (usableInfo.diskInfo.equals(diskInfo)) {
-                allocation = usableInfo.usableSlots;
-              }
-            }
-          }
-          sb.append(workerHost)
-              .append("-")
-              .append(diskInfo.mountPoint())
-              .append(" flushtime:")
-              .append(diskInfo.avgFlushTime())
-              .append(" fetchtime:")
-              .append(diskInfo.avgFetchTime())
-              .append(" allocation: ")
-              .append(allocation)
-              .append(" ");
-        }
-        sb.append(" | ");
-      }
-      logger.debug(
-          "total {} allocate with group {} with allocations {}",
-          partitionCnt,
-          StringUtils.join(groupAllocations, ','),
-          sb);
-    }
-    return restrictions;
   }
 
   private static PartitionLocation createLocation(
